+++ src/components/Button/Button.stories.tsx
@@ -1,5 +1,5 @@
 import { Story } from '@storybook/react'
-import React from 'react'
+import React, { useState } from 'react'
 import { action } from '@storybook/addon-actions'
 import styled, { css } from 'styled-components'
 
@@ -54,6 +54,25 @@ export const _Button: Story = () => {
               ボタン
             </Button>
             <Button variant="primary" disabled square onClick={action('clicked')}>
+              <FaPlusCircleIcon alt="プラスボタン" />
+            </Button>
+          </Cluster>
+          <Cluster>
+            <Button variant="primary" loading onClick={action('clicked')}>
+              ボタン
+            </Button>
+            <Button variant="primary" loading prefix={<FaPlusIcon />} onClick={action('clicked')}>
+              ボタン
+            </Button>
+            <Button
+              variant="primary"
+              loading
+              suffix={<FaPlusSquareIcon />}
+              onClick={action('clicked')}
+            >
+              ボタン
+            </Button>
+            <Button variant="primary" loading square onClick={action('clicked')}>
               <FaPlusCircleIcon alt="プラスボタン" />
             </Button>
           </Cluster>
@@ -118,6 +137,14 @@ export const _Button: Story = () => {
           </Cluster>
           <Cluster>
             <Button variant="primary" disabled size="s" onClick={action('clicked')}>
+              ボタン
+            </Button>
+          </Cluster>
+          <Cluster>
+            <Button variant="primary" disabled size="s" onClick={action('clicked')} loading>
+              ボタン
+            </Button>
+            <Button variant="primary" square disabled size="s" onClick={action('clicked')} loading>
               ボタン
             </Button>
           </Cluster>
@@ -321,6 +348,20 @@ _ButtonAnchorControl.argTypes = {
   children: { control: 'text', defaultValue: 'ボタン' },
   prefix: { control: 'text' },
   suffix: { control: 'text' },
+}
+
+export const WithLoading: Story = (args: ButtonProps) => {
+  const [loading, setLoading] = useState(false)
+  const handleClick = () => {
+    setLoading(true)
+    setTimeout(() => setLoading(false), 1000)
+  }
+
+  return (
+    <Button {...args} prefix={<FaPlusCircleIcon />} loading={loading} onClick={handleClick}>
+      読み込み
+    </Button>
+  )
 }
 
 const Wrapper = styled.div`