+++ src/components/Button/Button.tsx
@@ -1,9 +1,12 @@
 import React, { ButtonHTMLAttributes, forwardRef } from 'react'
+import styled, { css } from 'styled-components'
 
-import { BaseProps } from './types'
+import { Theme, useTheme } from '../../hooks/useTheme'
+import { BaseProps, Variant } from './types'
 import { useClassNames } from './useClassNames'
 import { ButtonWrapper } from './ButtonWrapper'
 import { ButtonInner } from './ButtonInner'
+import { Loader as shrLoader } from '../Loader'
 
 type ElementProps = Omit<ButtonHTMLAttributes<HTMLButtonElement>, keyof BaseProps>
 
@@ -17,13 +20,22 @@ export const Button = forwardRef<HTMLButtonElement, BaseProps & ElementProps>(
       suffix,
       wide = false,
       variant = 'secondary',
+      disabled,
       className = '',
       children,
+      loading = false,
       ...props
     },
     ref,
   ) => {
+    const theme = useTheme()
     const classNames = useClassNames().button
+
+    const loader = <Loader size="s" type="light" variant={variant} themes={theme} />
+    const actualPrefix = !loading && prefix
+    const actualSuffix = loading && !square ? loader : suffix
+    const disabledOnLoading = loading || disabled
+    const actualChildren = loading && square ? loader : children
 
     return (
       <ButtonWrapper
@@ -35,9 +47,11 @@ export const Button = forwardRef<HTMLButtonElement, BaseProps & ElementProps>(
         variant={variant}
         className={`${className} ${classNames.wrapper}`}
         buttonRef={ref}
+        disabled={disabledOnLoading}
+        loading={loading}
       >
-        <ButtonInner prefix={prefix} suffix={suffix}>
-          {children}
+        <ButtonInner prefix={actualPrefix} suffix={actualSuffix}>
+          {actualChildren}
         </ButtonInner>
       </ButtonWrapper>
     )
@@ -45,3 +59,22 @@ export const Button = forwardRef<HTMLButtonElement, BaseProps & ElementProps>(
 )
 // BottomFixedArea での判定に用いるために displayName を明示的に設定する
 Button.displayName = 'Button'
+
+const Loader = styled(shrLoader)<{ variant: Variant; themes: Theme }>`
+  ${({ variant, themes: { color } }) => css`
+    vertical-align: bottom;
+
+    &&& {
+      .s {
+        width: 1em;
+        height: 1em;
+      }
+    }
+
+    .light {
+      border-color: ${variant === 'secondary'
+        ? color.TEXT_DISABLED
+        : color.disableColor(color.TEXT_WHITE)};
+    }
+  `}
+`