+++ src/components/Loader/Loader.stories.tsx
@@ -1,4 +1,3 @@
-import { Theme, useTheme } from '../../hooks/useTheme'
 import { Story } from '@storybook/react'
 import * as React from 'react'
 import styled, { css } from 'styled-components'
@@ -11,8 +10,6 @@ export default {
 }
 
 export const All: Story = () => {
-  const theme = useTheme()
-
   return (
     <>
       <Wrapper>
@@ -33,7 +30,7 @@ export const All: Story = () => {
         </List>
       </Wrapper>
 
-      <GrayWrapper themes={theme}>
+      <GrayWrapper>
         <Text>Light</Text>
         <List>
           <dt>Default</dt>
@@ -54,24 +51,21 @@ export const All: Story = () => {
   )
 }
 All.storyName = 'all'
+All.parameters = { withTheming: true }
 
 const Wrapper = styled.div`
   padding: 24px;
 `
-const GrayWrapper = styled(Wrapper)<{ themes: Theme }>`
-  ${({ themes }) => {
-    const { color } = themes
+const GrayWrapper = styled(Wrapper)(
+  ({ theme: { color } }) => css`
+    background-color: ${color.SCRIM};
 
-    return css`
-      background-color: ${color.SCRIM};
-
-      p,
-      dt {
-        color: #fff;
-      }
-    `
-  }}
-`
+    p,
+    dt {
+      color: #fff;
+    }
+  `,
+)
 
 const Text = styled.p`
   margin: 0 0 16px;